Fix: pictureGift
세부 주석 추가 및 이미지 데이터 add 함수 추가
