--- liveOn/liveOn/ViewModels/ViewModel.swift
@@ -1,76 +1,0 @@
-//
-//  ViewModel.swift
-//  liveOn
-//
-//  Created by 이성민 on 2022/06/04.
-//
-
-import SwiftUI
-
-class ViewModel: ObservableObject {
-    @Published var image: UIImage?  // 그냥 이미지랑 UI 이미지 차이는 뭘까??
-    @Published var showPicker = false
-    @Published var source: Picker.Source = .library
-    @Published var showCameraAlert = false
-    @Published var cameraError: Picker.CameraErrorType?
-    
-    func showPhotoPicker() {
-        do {
-            if source == .camera {
-                try Picker.checkPermissions()
-            }
-            showPicker = true
-        } catch {
-            showCameraAlert = true
-            cameraError = Picker.CameraErrorType(error: error as! Picker.PickerError)
-        }
-    }
-}
-
-extension Color {
-    static let mainBackGround = Color(hex: "#FCFAFA")
-    static let primaryColor = Color(hex: "#A5726F")
-    
-    init(hex: String) {
-        let scanner = Scanner(string: hex)
-        _ = scanner.scanString("#")
-        
-        var rgb: UInt64 = 0
-        scanner.scanHexInt64(&rgb)
-        
-        let r = Double((rgb >> 16) & 0xFF) / 255.0
-        let g = Double((rgb >> 8) & 0xFF) / 255.0
-        let b = Double((rgb >> 0) & 0xFF) / 255.0
-        self.init(red: r, green: g, blue: b)
-    }
-}
-
-import Combine
-import SwiftUI
-
-final class KeyboardHandler: ObservableObject {
-    @Published private(set) var keyboardHeight: CGFloat = 0
-    
-    private var cancellabel: AnyCancellable?
-    
-    private let keyboardWillShow = NotificationCenter.default
-        .publisher(for: UIResponder.keyboardWillShowNotification)
-        .compactMap { ($0.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? CGRect)?.height }
-    
-    private let keyboardWillHide = NotificationCenter.default
-        .publisher(for: UIResponder.keyboardWillHideNotification)
-        .map { _ in CGFloat.zero }
-    
-    init() {
-        cancellabel = Publishers.Merge(keyboardWillShow, keyboardWillHide)
-            .subscribe(on: DispatchQueue.main)
-            .assign(to: \.self.keyboardHeight, on: self)
-    }
-}
-
-extension View {
-    func hideKeyboard() {
-        let resign = #selector(UIResponder.resignFirstResponder)
-        UIApplication.shared.sendAction(resign, to: nil, from: nil, for: nil)
-    }
-}
