+++ liveOn/liveOn/ViewModels/imageViewModel.swift
@@ -1,0 +1,100 @@
+
+import SwiftUI
+
+class imageViewModel: ObservableObject {
+    @Published var image: UIImage?
+    @Published var showPicker = false
+    @Published var source: Picker.Source = .library
+    @Published var showCameraAlert = false
+    @Published var cameraError: Picker.CameraErrorType?
+    @Published var imageName: String = ""
+    @Published var isEditing = false
+    @Published var selectedImage: MyImage?
+    @Published var myImages: [MyImage] = []
+    @Published var showFileAlert = false
+    @Published var appError: MyImageError.ErrorType?
+    
+    // MARK: 파일저장 시 경로를 프린트해줌
+    init() {
+        print(FileManager.docDirURL.path)
+    }
+    
+    func showPhotoPicker() {
+        do {
+            if source == .camera {
+                try Picker.checkPermissions()
+            }
+            showPicker = true
+        } catch {
+            showCameraAlert = true
+            cameraError = Picker.CameraErrorType(error: error as! Picker.PickerError)
+        }
+    }
+    
+    func reset() {
+        image = nil
+        imageName = ""
+    }
+    
+    // MARK: 현재 myImage 배열에 추가하는 함수지만, 추후 서버에 추가하는 기능으로 수정예정
+    func addMyImage(_ name: String, image: UIImage) {
+        reset()
+        let myImage = MyImage(name: name)
+        do {
+            try FileManager().saveImage("\(myImage.id)", image: image)
+            myImages.append(myImage)
+            saveMyImagesJSONFile()
+        } catch {
+            showFileAlert = true
+            appError = MyImageError.ErrorType(error: error as! MyImageError)
+        }
+    }
+    
+    func saveMyImagesJSONFile() {
+        let encoder = JSONEncoder()
+        do {
+            let data = try encoder.encode(myImages)
+            let jsonString = String(decoding: data, as: UTF8.self)
+            do {
+                try FileManager().saveDocument(contents: jsonString)
+            } catch {
+                showFileAlert = true
+                appError = MyImageError.ErrorType(error: error as! MyImageError)
+            }
+        } catch {
+            showFileAlert = true
+            appError = MyImageError.ErrorType(error: .encodingError)
+        }
+    }
+}
+
+
+import Combine
+import SwiftUI
+
+final class KeyboardHandler: ObservableObject {
+    @Published private(set) var keyboardHeight: CGFloat = 0
+    
+    private var cancellabel: AnyCancellable?
+    
+    private let keyboardWillShow = NotificationCenter.default
+        .publisher(for: UIResponder.keyboardWillShowNotification)
+        .compactMap { ($0.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? CGRect)?.height }
+    
+    private let keyboardWillHide = NotificationCenter.default
+        .publisher(for: UIResponder.keyboardWillHideNotification)
+        .map { _ in CGFloat.zero }
+    
+    init() {
+        cancellabel = Publishers.Merge(keyboardWillShow, keyboardWillHide)
+            .subscribe(on: DispatchQueue.main)
+            .assign(to: \.self.keyboardHeight, on: self)
+    }
+}
+
+extension View {
+    func hideKeyboard() {
+        let resign = #selector(UIResponder.resignFirstResponder)
+        UIApplication.shared.sendAction(resign, to: nil, from: nil, for: nil)
+    }
+}
