+++ liveOn/liveOn/Views/Letter/TextLimiter.swift
@@ -17,13 +17,16 @@ class TextLimiter: ObservableObject {
         value = placeholder
     }
     
-    @Published var value = "" {
+    @Published var value = ""
+    @Published var hasReachedLimit = false
+    @Published var inputEntered = false
+    {
         didSet {
             // 사용자가 입력했는지 체크
             if value.count > 0 &&  value != placeholder {
                 self.inputEntered = true } else {
-                self.inputEntered = false
-            }
+                    self.inputEntered = false
+                }
             
             // 제한글자를 넘어서면 더 이상 쳐지지 않게 하기
             if value.count > self.limit {
@@ -34,6 +37,4 @@ class TextLimiter: ObservableObject {
             }
         }
     }
-    @Published var hasReachedLimit = false
-    @Published var inputEntered = false
 }
